feat: remove compilation command
After thinking for some time, I've decided to get rid of compilation
command. This would be useful to compile container and plugin data at
build-time, but for the container the task is not trivial.

Compilation would require WordPress runtime, so the only viable solution
would be to boot whole WordPress with database. Removing the use of
WordPress functions wouldn't be possible or at least required some
common WP functions to be rewritten in our codebase.

For current approach, I bias towards compilation in customer's
environment at first runtime, with graceful handling of any IO operation
issues.

Signed-off-by: Bart Jaskulski <bjaskulski@protonmail.com>