feat: reintroduce command to compile plugin info
Actually, it's better to compile plugin info at build time, because
later, when client updates the plugin it may be hard to determine
whether cached file with plugin info should be
invalidated. Hypothetically, if client has plugin in version 1.0.0, and
updates to 1.0.1, there's nothing that could tell us, the plugin info
file is old. The only option would be to read the file and parse headers
every time, what would defy the cached file purpose. It's just easier
this way, especially when considering container compilation, which also
have to be version aware, to make sure only the recent version is used
for compilation.

Signed-off-by: Bart Jaskulski <bjaskulski@protonmail.com>