fix: improve the work of stoppable binder
Previously stoppable binder was working for all binders, although it
only supported to signal stop execution with class-based binders. This
changes the behavior, as having all binders wrapped in stoppable binder
resulted in execution of all collection binders, without regard for the
stop signal.

This is convoluted and gets even worse, so this part of the API should
be rewritten. The side effect of this change is that any callable based
binder (e.g. database migrations) will not be stopped, no matter what
happens. This is a serious quirk, and may be better to remove other
kinds of binders than hookable or rethink the stoppable binder.

Signed-off-by: Bart Jaskulski <bjaskulski@protonmail.com>